requirements.txt
streamlit==1.24.1
pandas>=2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.2
Pillow>=9.5
uvicorn
python-dotenv